+++ src/plugin.rs
@@ -1,7 +1,15 @@
 //! Contains main plugin exported by this crate.
 
+use crate::action_state::{ActionData, ActionState, Timing};
+use crate::axislike::{
+    AxisType, DeadZoneShape, DualAxis, DualAxisData, MouseMotionAxisType, MouseWheelAxisType,
+    SingleAxis, VirtualAxis, VirtualDPad,
+};
+use crate::buttonlike::{MouseMotionDirection, MouseWheelDirection};
 use crate::clashing_inputs::ClashStrategy;
-use crate::prelude::ActionState;
+use crate::dynamic_action::DynAction;
+use crate::input_map::InputMap;
+use crate::user_input::{InputKind, Modifier, UserInput};
 use crate::Actionlike;
 use core::hash::Hash;
 use core::marker::PhantomData;
@@ -9,7 +17,7 @@ use std::fmt::Debug;
 
 use bevy::app::{App, Plugin};
 use bevy::ecs::prelude::*;
-use bevy::input::InputSystem;
+use bevy::input::{ButtonState, InputSystem};
 use bevy::prelude::{PostUpdate, PreUpdate};
 use bevy::reflect::TypePath;
 #[cfg(feature = "ui")]
@@ -151,6 +159,26 @@ impl<A: Actionlike + TypePath> Plugin for InputManagerPlugin<A> {
         };
 
         app.register_type::<ActionState<A>>()
+            .register_type::<InputMap<A>>()
+            .register_type::<UserInput>()
+            .register_type::<InputKind>()
+            .register_type::<ActionData>()
+            .register_type::<Modifier>()
+            .register_type::<ActionState<A>>()
+            .register_type::<Timing>()
+            .register_type::<VirtualDPad>()
+            .register_type::<VirtualAxis>()
+            .register_type::<SingleAxis>()
+            .register_type::<DualAxis>()
+            .register_type::<AxisType>()
+            .register_type::<MouseWheelAxisType>()
+            .register_type::<MouseMotionAxisType>()
+            .register_type::<DualAxisData>()
+            .register_type::<DeadZoneShape>()
+            .register_type::<ButtonState>()
+            .register_type::<MouseWheelDirection>()
+            .register_type::<MouseMotionDirection>()
+            .register_type::<DynAction>()
             // Resources
             .init_resource::<ToggleActions<A>>()
             .init_resource::<ClashStrategy>();